Makes external organs compatible with color_src and hasinner, and gives inner overlays their own color_src
Pretty much what it says on the tin. This adds inner_color_src, which works like color_src but for the hasinner overlays, and also makes it so that external organs such as snouts will be properly colored by color_src and can have hasinner overlays. This doesn't affect anything presently in the game, but it will be very useful for adding more species features in the future, such as snouts with differently-colored noses attached to them, or tails with two different colors.